# This is a combination of 7 commits. # This is the 1st commit message:
feat: add multiple public APIs, improve README structure and fix minor issues

- Added OpenAI, UIFaces, UINames, RandomUser, TFL, iRail, German DB, Netflix Roulette
- Included Swiss, Norwegian, Budapest, and Toronto (TTC) transport APIs
- Added OpenWeatherMap, Apple iTunes, Spotify, Open Notify, Adorable Avatars
- Added Ron Swanson Quotes, Twitch, Comic Vine, GitHub User, and Temp Email APIs
- Fixed broken NASA link and spelling issues (e.g. Dribble to Dribbble)
- Corrected Dutch translations and organized APIs into clearer categories
- Fixed formatting (eg: wrong column)
- Managed to restore TempMail
- Removed TempMail

# This is the commit message #2:

Add JSONPlaceholder
# This is the commit message #3:

Fedger.io 

This is a publicly available API comprising of several sub-APIs: company, fundings, geodata, news etc. 
The API itself is proprietary as the company behind charges for its usage but there's a certain amounts of "free calls" for every registered consumer.
Registering the API key is very easy and for free. I have also written a PureScript library to access the API (the backend logic is written in JS, so it can be easily used without PureScript as well).

Not sure if such APIs are welcome here, so feel free to reject this PR.
# This is the commit message #4:

Fedger.io 

This is a publicly available API comprising of several sub-APIs: company, fundings, geodata, news etc. 
The API itself is proprietary as the company behind charges for its usage but there's a certain amounts of "free calls" for every registered consumer.
Registering the API key is very easy and for free. I have also written a PureScript library to access the API (the backend logic is written in JS, so it can be easily used without PureScript as well).

Not sure if such APIs are welcome here, so feel free to reject this PR.
# This is the commit message #5:

adding Wordnik dictionary data API

# This is the commit message #6:

Update README.md
# This is the commit message #7:

chore: Add FitBit API