Update README.md to break down Media category
The Media category was becoming too vague and too cluttered with items that fell into other, already existing, categories. By breaking up this category, we are able to allow viewers to quickly find what they need, instead of making them search the category they want and then have to double-check the "catch-all" basin that was Media.