Remove Okta
As discussed in PR #407, Stormpath closed and joined with Okta. However, there is no indication of a free tier for Okta now. The closest thing I could find on their pricing page was joining 1 service to an Okta SSO for free, but that isn't any form of an API that we cover in this project.